docs, tests, new ExceptionFormatter
